data now is reassigned instead of adding each process to it in order, many things deleted
there was strange occurrences with old model. The same pid was appearing 2, 3, 4, 5, 6 times!!! I got tired on trying to solve that. With this, also the manual ordering and separate process to update visible cells are no more needed.
Also in this commit: create a timer to use as decorator and solves a type warning with explicit variable type declaration