better method of showing selections as text & bug corrections
instead of 'but' to each exception group inside a 'all', which means one of the 'all's will be redundant or confusing, now the exceptions are grouped at one 'except' at the end of the text.

there was a bug on select_rows(True) bcs a data was being processed even after being removed of self.selection_control, bcs the iteration is over a copy.

also, L447 was interpreting a data with _check == False as a search selection (?). solved with L458