performance improvements & more
`processes` keys and "proc_pid" item of each `rv.data` item are no longer a str but the int it should be.
No need for converting each pid to string on each `update_processes` call.

For same reasons, "proc_cpu" and "proc_mem" items of each `rv.data` are no longer str but float.
So now, `main.key_func` do not have to convert and change strings in each call. The string used by GUI is just derived from the real value, instead of being stored like that

`process_iter` calls no longer ask for "pid", as it is always stored in a class variable. So `proc.info["pid"]` is now `proc.pid`.

On `update_processes` there was two iterations over the results of `process_iter`, now is just a faster one.

In the same function the selection label was being updated wrongly. Just the first selection control found with the inexistent pid would be deleted, including an entire search. Now it calls the same as if the checkbox was unmarked.

For first ordering, `order` was being called to set `key_func`, `order_by` and `reverse`. Now its made on `__init__`

`correct_special_order_cell` and `correct_order_cell` had a assign of a modified dict into the "old dict", but the modified dict is just a shallow copy, so the assign was not needed at all

Old `existent_search` is the same as checking for `search` bcs for Python empty strings return False

`fast_search` and other methods which need to put Threads in lists were appending and then accessing the last element. Now they set a local variable, append the local variable and keep with acess to the Thread without entering the list.

`fast_search` and `select_row` were considering a no lowercased text

Repeat of same iteration to find variables and methods in `set_visible_range` solved

`sorted_by` variable was not being used

`title` local variable of `show_fails` now more readable (and maybe faster)

unneed `not` operator in two ìf else` at main.kv removed

:face_with_head_bandage: